+++ include/raccoon/internals/challenge.hpp
