+++ include/raccoon/internals/masked_poly.hpp
