+++ include/raccoon/internals/poly.hpp
