+++ include/raccoon/internals/poly_mat.hpp
