+++ include/raccoon/internals/poly_vec.hpp
