+++ include/raccoon/internals/serialization.hpp
