+++ include/raccoon/internals/utils.hpp
